requirements.txt
pandas==2.2.2
transformers==4.43.3
torch>=2.2.0
requests==2.32.3
python-dotenv==1.0.1
tqdm==4.66.4
streamlit==1.37.1
uvicorn==0.30.1
fastapi==0.111.0
